[REFACTOR] 레거시 v1 Auth 관련 invite페이지 삭제
