+++ src/App.js
@@ -1,7 +1,6 @@
 import React from "react";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes, Navigate } from "react-router-dom";
 import "./App.css";
-import Invite from "./features/invite/pages/Invite";
 import EventCalendar from "./features/calendar/pages/EventCalendar";
 import CreateCalendar from "./features/calendar/pages/CreateCalendar";
 import GetAppointmentRedirect from "./features/calendar/pages/GetAppointmentRedirect";
@@ -14,6 +13,18 @@ import Login from "./features/auth/pages/Login";
 import Register from "./features/auth/pages/Register";
 import KakaoCallback from "./features/auth/pages/KakaoCallback";
 
+// Legacy invite route redirect component
+const InviteRedirect = () => {
+  const urlParams = new URLSearchParams(window.location.search);
+  const appointmentId = urlParams.get("appointmentId");
+  
+  if (appointmentId) {
+    return <Navigate to={`/getAppointment?appointmentId=${appointmentId}`} replace />;
+  }
+  
+  return <Navigate to="/" replace />;
+};
+
 function App() {
   return (
     <Router>
@@ -22,7 +33,7 @@ function App() {
           <Route path="/" element={<LandingPage />} />
           <Route path="/MonthView" element={<CreateCalendar />} />
           <Route path="/getAppointment" element={<GetAppointmentRedirect />} />
-          <Route path="/invite" element={<Invite />} />
+          <Route path="/invite" element={<InviteRedirect />} />
           <Route path="/menu" element={<Menu />} />
 
           <Route path="/profile/edit" element={<EditProfile />} />
