+++ src/features/auth/pages/Login.js
@@ -181,21 +181,14 @@ const Login = () => {
 
           setResponseMessage("로그인 성공!");
 
-          // 리다이렉트 URL이 있고 invite 페이지면 개인 캘린더로, 아니면 원래 URL 또는 MonthView로 이동
+          // 리다이렉트 URL이 있고 getAppointment 페이지면 그대로 리다이렉트, 아니면 원래 URL 또는 MonthView로 이동
           const redirectUrl = searchParams.get("redirect");
           setTimeout(() => {
             if (redirectUrl) {
               const decodedUrl = decodeURIComponent(redirectUrl);
-              // invite 페이지에서 온 경우 개인 캘린더로 이동
-              if (decodedUrl.includes("/invite?appointmentId=")) {
-                const urlParams = new URLSearchParams(decodedUrl.split('?')[1]);
-                const appointmentId = urlParams.get("appointmentId");
-                navigate("/individualCalendar", {
-                  state: { 
-                    appointmentId: appointmentId,
-                    userName: "참여자"
-                  }
-                });
+              // getAppointment 페이지에서 온 경우 그대로 리다이렉트
+              if (decodedUrl.includes("/getAppointment?appointmentId=")) {
+                navigate(decodedUrl);
               } else {
                 navigate(decodedUrl);
               }
