+++ src/store/index.ts
@@ -1,4 +1,3 @@
 export { useUserStore } from "./userStore";
 export { useUIStore } from "./uiStore";
 export { useCalendarStore } from "./calendarStore";
-export { useInviteStore } from "./inviteStore";
